+++ Bloch/Cards/FocusCard.cs
@@ -6,11 +6,11 @@ using System.Reflection;
 
 namespace Shockah.Bloch;
 
-internal sealed class FocusCard : Card, IRegisterable
+internal sealed class FocusCard : Card, IRegisterable, IHasCustomCardTraits
 {
 	public static void Register(IPluginPackage<IModManifest> package, IModHelper helper)
 	{
-		helper.Content.Cards.RegisterCard(MethodBase.GetCurrentMethod()!.DeclaringType!.Name, new()
+		var entry = helper.Content.Cards.RegisterCard(MethodBase.GetCurrentMethod()!.DeclaringType!.Name, new()
 		{
 			CardType = MethodBase.GetCurrentMethod()!.DeclaringType!,
 			Meta = new()
@@ -22,13 +22,19 @@ internal sealed class FocusCard : Card, IRegisterable
 			Art = helper.Content.Sprites.RegisterSpriteOrDefault(package.PackageRoot.GetRelativeFile("assets/Cards/Focus.png"), StableSpr.cards_BigShield).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Focus", "name"]).Localize
 		});
+		
+		ModEntry.Instance.KokoroApi.Finite.SetBaseFiniteUses(entry.UniqueName, Upgrade.None, 3);
+		ModEntry.Instance.KokoroApi.Finite.SetBaseFiniteUses(entry.UniqueName, Upgrade.A, 3);
 	}
 
 	public override CardData GetData(State state)
-		=> new()
+		=> new() { cost = 1 };
+
+	public IReadOnlySet<ICardTraitEntry> GetInnateTraits(State state)
+		=> upgrade switch
 		{
-			cost = 1,
-			infinite = upgrade != Upgrade.B
+			Upgrade.B => new HashSet<ICardTraitEntry>(),
+			_ => new HashSet<ICardTraitEntry> { ModEntry.Instance.KokoroApi.Finite.Trait },
 		};
 
 	public override List<CardAction> GetActions(State s, Combat c)
@@ -44,16 +50,14 @@ internal sealed class FocusCard : Card, IRegisterable
 					_ => 1
 				}
 			},
-			new AStatus
-			{
-				targetPlayer = true,
-				status = AuraManager.VeilingStatus.Status,
-				statusAmount = upgrade switch
+			ModEntry.Instance.Api.MakeChooseAura(
+				card: this,
+				amount: upgrade switch
 				{
 					Upgrade.A => 2,
 					Upgrade.B => 3,
-					_ => 1
+					_ => 1,
 				}
-			}
+			),
 		];
 }